Move price conversion check to a filter
